requirements.txt
torch==2.1.0
transformers==4.40.0
panda
nltk
gradio
seaborn
matplotlib